move std to func
