+++ README.md
@@ -1,7 +1,5 @@
 # xDSL
 
-TODO
-
 ## Prerequisits
 
 To install all required dependencies, execute the following command:
@@ -23,6 +21,17 @@ pytest
 # Executes filecheck tests
 lit tests/filecheck
 ```
+
+## Generating executables through MLIR
+
+xDSL can generate executables using MLIR as the backend. To use this functionality, make sure to install the [MLIR Python Bindings](https://mlir.llvm.org/docs/Bindings/Python/). Given an input file `input.xdsl`, that contains IR with only the mirrored dialects found in `src/xdsl/dialects` (arith, memref, func, cf, scf, and builtin), run:
+
+```
+mlir-opt --convert-scf-to-cf --convert-cf-to-llvm --convert-func-to-llvm --convert-arith-to-llvm --convert-memref-to-llvm --reconcile-unrealized-casts *input.xdsl* | mlir-translate --mlir-to-llvmir > tmp.ll
+```
+
+The generated `tmp.ll` file contains LLVMIR, so it can be directly passed to a compiler like clang.
+
 
 ## Formatting
