+++ tests/mlir_converter_test.py
@@ -3,7 +3,7 @@ import pytest
 docutils = pytest.importorskip("mlir")
 from xdsl.mlir_converter import *
 from xdsl.dialects.scf import Scf
-from xdsl.dialects.std import Std
+from xdsl.dialects.func import Func
 from xdsl.dialects.memref import MemRef
 from xdsl.dialects.affine import Affine
 from xdsl.dialects.arith import Arith
@@ -13,7 +13,7 @@ from xdsl.parser import Parser
 def convert_and_verify(test_prog: str):
     ctx = MLContext()
     builtin = Builtin(ctx)
-    std = Std(ctx)
+    func = Func(ctx)
     affine = Affine(ctx)
     arith = Arith(ctx)
     scf = Scf(ctx)
@@ -28,17 +28,18 @@ def convert_and_verify(test_prog: str):
         mlir_ctx.allow_unregistered_dialects = True
         with mlir.ir.Location.unknown(mlir_ctx):
             mlir_prog = converter.convert_op(module)
+            print(mlir_prog)
             assert (mlir_prog.verify())
 
 
-def test_std_conversion():
+def test_func_conversion():
     test_prog = """
 module() {
-  builtin.func() ["sym_name" = "test", "type" = !fun<[], []>, "sym_visibility" = "private"]
+  func.func() ["sym_name" = "test", "function_type" = !fun<[], []>, "sym_visibility" = "private"]
   {
-    std.return()
+    func.return()
   }
-  std.call() ["callee" = !flat_symbol_ref<"test">]
+  func.call() ["callee" = !flat_symbol_ref<"test">]
 }
     """
     convert_and_verify(test_prog)
@@ -47,11 +48,11 @@ module() {
 def test_arith_conversion():
     test_prog = """
 module() {
-  builtin.func() ["sym_name" = "test", "type" = !fun<[!i32], [!i32]>, "sym_visibility" = "private"]{
+  func.func() ["sym_name" = "test", "function_type" = !fun<[!i32], [!i32]>, "sym_visibility" = "private"]{
   ^0(%arg : !i32):
     %0 : !i32 = arith.constant() ["value" = 0 : !i32]
     %res : !i32 = arith.addi(%arg : !i32, %0 : !i32) 
-    std.return(%res : !i32)
+    func.return(%res : !i32)
   }
 }
     """
@@ -61,7 +62,7 @@ module() {
 def test_scf_conversion():
     test_prog = """
 module() {
-  builtin.func() ["sym_name" = "test", "type" = !fun<[!i32], [!i32]>, "sym_visibility" = "private"]{
+  func.func() ["sym_name" = "test", "function_type" = !fun<[!i32], [!i32]>, "sym_visibility" = "private"]{
   ^0(%arg : !i32):
     %0 : !i32 = arith.constant() ["value" = 42 : !i32]
     %t : !i1 = arith.cmpi(%arg : !i32, %0 : !i32) ["predicate" = 1 : !i64]
@@ -72,7 +73,7 @@ module() {
         %else : !i32 = arith.subi(%arg : !i32, %0 : !i32)
         scf.yield(%else : !i32)
     }
-    std.return(%res : !i32)
+    func.return(%res : !i32)
   }
 }
     """
@@ -82,11 +83,11 @@ module() {
 def test_variadic_conversion():
     test_prog = """
 module() {
-  builtin.func() ["sym_name" = "test", "type" = !fun<[], []>, "sym_visibility" = "private"]
+  func.func() ["sym_name" = "test", "function_type" = !fun<[], []>, "sym_visibility" = "private"]
   {
     %1 : !memref<[1 : !i64], !i32> = memref.alloca() ["alignment" = 0 : !i64, "operand_segment_sizes" = !dense<!vector<[2 : !i64], !i32>, [0 : !i32, 0 : !i32]>]
 
-    std.return()
+    func.return()
   }
 }
     """
@@ -96,7 +97,7 @@ module() {
 def test_memref_conversion():
     test_prog = """
 module() {
-  builtin.func() ["sym_name" = "sum", "type" = !fun<[!i32, !i32], [!i32]>, "sym_visibility" = "public"]{
+  func.func() ["sym_name" = "sum", "function_type" = !fun<[!i32, !i32], [!i32]>, "sym_visibility" = "public"]{
   ^0(%0 : !i32, %1 : !i32):
     %2 : !index = arith.constant() ["value" = 0 : !index]
     %3 : !memref<[1 : !index], !i32> = memref.alloca() ["alignment" = 0 : !i64, "operand_segment_sizes" = !dense<!vector<[2 : !index], !i32>, [0 : !i32, 0 : !i32]>]
@@ -109,7 +110,7 @@ module() {
     %8 : !index = arith.constant() ["value" = 0 : !index]
     %9 : !i32 = memref.load(%5 : !memref<[1 : !index], !i32>, %8 : !index)
     %10 : !i32 = arith.addi(%7 : !i32, %9 : !i32)
-    std.return(%10 : !i32)
+    func.return(%10 : !i32)
   }
 }
     """
