+++ tests/use_test.py
@@ -1,7 +1,6 @@
 from xdsl.dialects.builtin import *
 from xdsl.parser import Parser
 from xdsl.printer import Printer
-from xdsl.dialects.std import *
 from xdsl.dialects.arith import *
 
 test_prog = """
